+++ lib/isodoc/rsd/metadata.rb
@@ -3,7 +3,7 @@ require "isodoc"
 module IsoDoc
   module Rsd
 
-    class Metadata < IsoDoc::Acme::Metadata
+    class Metadata < IsoDoc::Generic::Metadata
       def configuration
         Metanorma::Rsd.configuration
       end
